+++ src/main.ts
@@ -7,12 +7,6 @@ import { TimeoutInterceptor } from '@/common/interceptors/timeout.interceptor';
 import { swaggerConfig } from '@/docs/swagger-config';
 import { HttpExceptionFilter } from '@/common/filters/http-exception.filter';
 
-const key = 'fetch';
-
-global[key] = require('node-fetch');
-
-declare const module: any;
-
 async function bootstrap(): Promise<void> {
   const logger = new Logger('Main');
 
@@ -32,10 +26,5 @@ async function bootstrap(): Promise<void> {
   });
 
   await app.listen(process.env.PORT, () => logger.log(`App running 🔥`));
-
-  if (module.hot) {
-    module.hot.accept();
-    module.hot.dispose(() => app.close());
-  }
 }
 bootstrap();